requirements.txt
requests~=2.27.1
python-dotenv~=0.20.0
beautifulsoup4~=4.11.1
oauthlib~=3.2.0